requirements.txt
psycopg2==2.7.1
pytest==3.1.0
testing.postgresql==1.3.0
